changed the structure of folder and convert all to pure components
