+++ tsconfig.json
@@ -22,7 +22,8 @@
   "include": [
     "next-env.d.ts",
     "**/*.ts",
-    "**/*.tsx"
+    "**/*.tsx",
+    "components/sidebar/chart.jsx"
   ],
   "exclude": [
     "node_modules"
